requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
cloudinary
python-dotenv
langchain
openai
faiss-cpu
